requirements.txt
spotipy>=2.23.0
requests>=2.31.0
shazamio>=0.8.0
pyyaml>=6.0
